refactor: partial ahl support, notification rework
notifications now use the playbyplay endpoint for stricter play definitions. we also ditch D1 for "live" game cache and use KV instead. for posterity, i primarily did this because i was getting "too many SQL variables" when upserting 15 values and didn't want to deal with it - but I prefer this implementation anyway.